made editTargetVerseSource.js a helper
