+++ src/js/actions/VerseEditActions.js
@@ -1,7 +1,6 @@
 import consts from '../actions/CoreActionConsts';
 import {generateTimestamp} from '../helpers/index';
-import fs from 'fs-extra';
-import path from 'path-extra';
+import {editTargetVerseSource} from '../helpers/editTargetVerseSource';
 
 /**
  * @description This action updates or adds the data needed to
@@ -23,8 +22,8 @@ export const addVerseEdit = (before, after, tags, userName) => {
       userName,
       modifiedTimestamp: generateTimestamp()
     });
+    editTargetVerseSource(state, after);
     dispatch(editTargetVerseInBiblesReducer(after));
-    dispatch(editTargetVerseSource());
     dispatch({
       type: consts.TOGGLE_VERSE_EDITS_IN_GROUPDATA,
       contextId
@@ -48,42 +47,5 @@ export function editTargetVerseInBiblesReducer(editedText) {
       chapter,
       verse
     });
-  });
-}
-
-
-export function editTargetVerseSource() {
-  return ((dispatch, getState) => {
-    let {contextIdReducer, projectDetailsReducer, verseEditReducer} = getState();
-    let saveLocation = projectDetailsReducer.projectSaveLocation;
-    let chapter = contextIdReducer.contextId.reference.chapter; 
-    let verse = contextIdReducer.contextId.reference.verse;
-    if (chapter < 10) {
-      chapter = "0" + chapter;
-    }
-    let writeFolder = path.join(saveLocation, chapter);
-    let contents = fs.readdirSync(writeFolder);
-    let writeFile = null;
-    for (let i = 0; i < contents.length; i++) {
-      let chunk = contents[i];
-      let numChunk = parseInt(chunk.split('.txt')[0]);
-      if (numChunk <= verse && (contents[i + 1] === undefined || parseInt(contents[i + 1].split('.txt')[0]) > verse)) {
-        writeFile = path.join(writeFolder, chunk);
-        break;
-      }
-    }
-    let fileContents = fs.readFileSync(writeFile).toString();
-    let tokenizedFile = fileContents.split('\\v');
-    for (let i = 0; i < tokenizedFile.length; i++) {
-      let item = tokenizedFile[i];
-      let trimmedItem = item.trim();
-      let currentVerse = parseInt(trimmedItem.split(' ')[0]);
-      if (currentVerse === verse) {
-        tokenizedFile[i] = " " + verse + " " + verseEditReducer.after + " ";
-        break;
-      }
-    }
-    let saveFile = tokenizedFile.join('\\v');
-    fs.writeFileSync(writeFile, saveFile);
   });
 }
